Addressed various warnings from gcc (compiled using -Wall -W -Wcast-align) as part of making the code 64-bit clean.

+++ HTTPCache.h
@@ -161,7 +161,7 @@ public:
 	time_t age;
 	time_t max_age;		// From Cache-Control
 
-	unsigned long size;	// Size of cached entity body
+	unsigned int size;	// Size of cached entity body
 	bool range;		// Range is not currently supported. 10/02/02
 				// jhrg 
 
@@ -194,7 +194,7 @@ public:
     // they can access private stuff). We should not need this any longer,
     // but I'm hesitant to remove it since I cannot easily test with VC++.
     // 01/23/04 jhrg
-    unsigned long d_max_entry_size;	// Max individual entry size.
+    unsigned int d_max_entry_size;	// Max individual entry size.
 
     void remove_cache_entry(CacheEntry *entry) throw(InternalErr);
     bool stopGC() const;
@@ -211,13 +211,13 @@ private:
     bool d_expire_ignored;
     bool d_always_validate;
 
-    unsigned long d_total_size;	// How much can we store?
-    unsigned long d_folder_size; // How much of that is meta data?
-    unsigned long d_gc_buffer;	// How much memory needed as buffer?
+    unsigned int d_total_size;	// How much can we store?
+    unsigned int d_folder_size; // How much of that is meta data?
+    unsigned int d_gc_buffer;	// How much memory needed as buffer?
 #ifndef WIN32  //  Declared public above for win32
-    unsigned long d_max_entry_size;	// Max individual entry size.
+    unsigned int d_max_entry_size;	// Max individual entry size.
 #endif
-    unsigned long d_current_size;
+    unsigned int d_current_size;
     int d_default_expiration;
     unsigned int d_block_size;	// File block size.
 
@@ -351,11 +351,11 @@ public:
     void set_expire_ignored(bool mode);
     bool is_expire_ignored() const;
 
-    void set_max_size(unsigned long size);
-    unsigned long get_max_size() const;
+    void set_max_size(unsigned int size);
+    unsigned int get_max_size() const;
 
-    void set_max_entry_size(unsigned long size);
-    unsigned long get_max_entry_size() const;
+    void set_max_entry_size(unsigned int size);
+    unsigned int get_max_entry_size() const;
 
     void set_default_expiration(int exp_time);
     int get_default_expiration() const;
